Ver. 0.0.0.001

+++ R/read_scrapeBill.R
@@ -4,12 +4,14 @@
 #' 
 #' @param x A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.
 #' @param file A connection or the name of the file (use \code{.rds} extension).
+#' @param aux Whether to export auxiliary files into separate folder when \code{write_scrapeBill}. 
+#' Set to \code{TRUE} if you want to avoid the \code{.rds} file size to become too big.
 #' @param ... Additional argument passed to \code{\link[base]{saveRDS}}
 #' 
 #' @return For \code{read_scrapeBill}, an \code{R} object. 
-#' For \code{write_scrapeBill}, \code{NULL} invisibly. \code{write_scrapeBill} also writes out 
-#' auxiliary files under \code{FILENAME_files} folder. DON'T delete the folder. 
-#' \code{read_scrapeBill} will need the files in the folder to read file.
+#' For \code{write_scrapeBill}, \code{NULL} invisibly. If \code{write_scrapeBill}
+#' writes out auxiliary files under \code{FILENAME_files} folder, DON'T delete the 
+#' folder. \code{read_scrapeBill} will need the files in the folder to read file.
 #' 
 #' @importFrom xml2 xml_serialize
 #' @importFrom xml2 xml_unserialize
@@ -20,19 +22,25 @@
 #' @export
 read_scrapeBill <- function(file) {
   
+  # Read File
+  x <- readRDS(file)
+
   filename <- basename(file_path_sans_ext(file))
   foldername <- paste0(filename,"_files")
   dirloc <- paste(dirname(file),foldername,sep="/")
-  if (dir.exists(dirloc)==FALSE) stop("auxiliary files not found!")
-  
-  # Read File
-  x <- readRDS(file)
   
   # Put xml object back
   xmlback <- function(obj) {
-    tf <- paste(dirloc,obj,sep="/")
-    con <- file(tf, "rb")
-    on.exit(close(con), add = TRUE)
+    if (class(obj)[1]=="character") {
+      if (dir.exists(dirloc)==FALSE) stop("auxiliary files not found!")
+      tf <- paste(dirloc,obj,sep="/")
+      con <- file(tf, "rb")
+      on.exit(close(con), add = TRUE)
+    } else if (class(obj)[1]=="raw"){
+      con <- obj
+    } else {
+      stop("invalid object type!")
+    }
     res <- xml_unserialize(con)
     return(res)
   }
@@ -49,7 +57,7 @@ read_scrapeBill <- function(file) {
   } else if (class(x)[1]=="list" & class(x[[1]])=="scrapeBill") {
     x <- pblapply(x, function(k) writexml(k))
   } else {
-    stop("invalid object type.")
+    stop("invalid object type!")
   }
   
   return(x)
@@ -58,29 +66,37 @@ read_scrapeBill <- function(file) {
 
 #' @rdname read_scrapeBill
 #' @export
-write_scrapeBill <- function(x, file, ...) {
+write_scrapeBill <- function(x, file, aux=FALSE,...) {
   
-  filename <- basename(file_path_sans_ext(file))
-  foldername <- paste0(filename,"_files")
-  dirloc <- paste(dirname(file),foldername,sep="/")
-  if (dir.exists(dirloc)==TRUE) {
-    unlink(paste0(dirloc,"/*"))
-  } else {
-    dir.create(dirloc)
+  if (aux==TRUE) {
+    filename <- basename(file_path_sans_ext(file))
+    foldername <- paste0(filename,"_files")
+    dirloc <- paste(dirname(file),foldername,sep="/")
+    print(dir.exists(dirloc))
+    if (dir.exists(dirloc)==TRUE) {
+      unlink(paste0(dirloc,"/*"))
+    } else {
+      dir.create(dirloc)
+    }
   }
 
-  locback <- function(obj) {
-    obj <- html_nodes(obj, xpath="/html")
-    tf <- tempfile(tmpdir=dirloc)
-    con <- file(tf, "wb")
-    xml_serialize(obj, con)
-    on.exit(close(con), add = TRUE)
-    return(basename(tf))
+  locback <- function(obj, aux) {
+    if (aux==TRUE) {
+      obj <- html_nodes(obj, xpath="/html")
+      tf <- tempfile(tmpdir=dirloc)
+      con <- file(tf, "wb")
+      on.exit(close(con), add = TRUE)
+      xml_serialize(obj, con)
+      res <- basename(tf)
+    } else {
+      res <- xml_serialize(obj, NULL)
+    }
+    return(res)
   }
   
   writeloc <- function(k) {
-    k$info <- locback(k$info)
-    k$text <- locback(k$text)
+    k$info <- locback(k$info, aux=aux)
+    k$text <- locback(k$text, aux=aux)
     class(k) <- "scrapeBill"
     return(k)
   }
